requirements.txt
# Core Dependencies
python-dotenv>=0.19.0
pandas>=1.3.0
schedule>=1.1.0
SQLAlchemy>=1.4.0                       # Database abstraction layer

# Database Drivers
pymysql>=1.0.0                          # MySQL support
psycopg2-binary>=2.9.0                  # PostgreSQL support

# Optional: For SQL Server support
# pyodbc>=4.0.0

# Optional: Advanced scheduling
APScheduler>=3.9.0

# Optional: Email templates
jinja2>=3.0.0

# Testing
pytest>=7.0.0
pytest-mock>=3.6.0
